Sending process objects with a 'readline' method.
This is useful when you already have the value to be parsed out of a file. For
example, if you have a String, it can be converted to a StringIO and send to
process:

string = "hello there"
sio = StringIO.new(string)
SmarterCSV.process(sio)

There is no need to write the string on a new file, avoiding creating temporary
files, disk writes and so on.