Remove trailing slash from link to Jack article
The readme link to the article by Jack throws a "Sorry this page does not exist =(" error when including a trailing slash. Removing the slash fixes the problem. 

Found the issue by clicking the link in the readme and getting an error page. Googled for the article and found it. Compared the two urls and found only variance was the trailing slash. Tried it both ways in Chrome and Safari and it resolves the issue.