Refactor `guess_column_separator` logic
I found the original method's logic too prone to
fail even with simple scenarios. E.g., given this
CSV

```
id,description,created_at,updated_at
1,Acquaforte,2022-10-04 16:00:47 UTC,2022-10-04 16:00:47 UTC
2,Calcografia,2022-10-04 16:00:47 UTC,2022-10-04 16:00:47 UTC
3,Cromolitografia,2022-10-04 16:00:47 UTC,2022-10-04 16:00:47 UTC
4,Cromotipografia,2022-10-04 16:00:47 UTC,2022-10-04 16:00:47 UTC
5,Eliotipia,2022-10-04 16:00:47 UTC,2022-10-04 16:00:47 UTC
```

the guessed separator would be a colon (`:`) because we have
only 4 columns separated by 3 commans (`,`), while in values
there are 2 date-time-ish strings with 4 total colons (`:`)
as hours/minutes/seconds separators.

My assumption here is that if we have headers in file, then
the headers alone are the most affordable source of truth
regarding the separator, thus we scan for possible separator
only in the first line.

If we haven't headers, then we fallback to the original
guessing algorithm.

This refactor turns the method from being flaky to be
more consistend and predictable. This means that it
**should be considered a breaking change** even if proposed
as a bugfix.